src/plugins/data/common/search/aggs/buckets/filter_fn.test.ts
/*
 * Licensed to Elasticsearch B.V. under one or more contributor
 * license agreements. See the NOTICE file distributed with
 * this work for additional information regarding copyright
 * ownership. Elasticsearch B.V. licenses this file to you under
 * the Apache License, Version 2.0 (the "License"); you may
 * not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import { functionWrapper } from '../test_helpers';
import { aggFilter } from './filter_fn';

describe('agg_expression_functions', () => {
  describe('aggFilter', () => {
    const fn = functionWrapper(aggFilter());

    test('fills in defaults when only required args are provided', () => {
      const actual = fn({});
      expect(actual).toMatchInlineSnapshot(`
        Object {
          "type": "agg_type",
          "value": Object {
            "enabled": true,
            "id": undefined,
            "params": Object {
              "customLabel": undefined,
              "geo_bounding_box": undefined,
              "json": undefined,
            },
            "schema": undefined,
            "type": "filter",
          },
        }
      `);
    });

    test('includes optional params when they are provided', () => {
      const actual = fn({
        geo_bounding_box: JSON.stringify({
          wkt: 'BBOX (-74.1, -71.12, 40.73, 40.01)',
        }),
      });

      expect(actual.value).toMatchInlineSnapshot(`
        Object {
          "enabled": true,
          "id": undefined,
          "params": Object {
            "customLabel": undefined,
            "geo_bounding_box": Object {
              "wkt": "BBOX (-74.1, -71.12, 40.73, 40.01)",
            },
            "json": undefined,
          },
          "schema": undefined,
          "type": "filter",
        }
      `);
    });

    test('correctly parses json string argument', () => {
      const actual = fn({
        json: '{ "foo": true }',
      });

      expect(actual.value.params.json).toEqual({ foo: true });

      expect(() => {
        fn({
          json: '/// intentionally malformed json ///',
        });
      }).toThrowErrorMatchingInlineSnapshot(`"Unable to parse json argument string"`);
    });
  });
});
